pyproject.toml
# Copyright 2021 Canonical Ltd.
# See LICENSE file for licensing details.
[project]
name = "blackbox-exporter-k8s"
version = "0.0"
requires-python = "~=3.8"

dependencies = [
  "ops",
  "pyyaml",
  "lightkube>=0.11",  # observability_libs
  "lightkube-models",  # observability_libs
  "jsonschema", # traefik_k8s, tls_certificates
  "cosl",
]

[project.optional-dependencies]
dev = [
  # Linting
  "ruff",
  "codespell",
  # Static
  "pyright",
  # Unit
  "pytest",
  "coverage[toml]",
  "deepdiff",
  "hypothesis",
  "validators>=0.21.2",
  "pydantic>2.0",
  "ops-scenario",
  # Interface
  "pytest-interface-tester>0.3",
  # Integration
  "juju",
  "websockets<14.0",
  "pytest-operator",
  "pytest-httpserver",
  "sh",
]

# Testing tools configuration
[tool.coverage.run]
branch = true

[tool.coverage.report]
show_missing = true

# Formatting tools configuration
[tool.black]
line-length = 99
target-version = ["py38"]

# Linting tools configuration
[tool.ruff]
line-length = 99
extend-exclude = ["__pycache__", "*.egg_info"]

[tool.ruff.lint]
select = ["E", "W", "F", "C", "N", "R", "D", "I001"]
# Ignore E501 because using black creates errors with this
# Ignore D107 Missing docstring in __init__
ignore = ["E501", "D107", "N818", "RET504"]
# D100, D101, D102, D103: Ignore missing docstrings in tests
per-file-ignores = {"tests/*" = ["D100","D101","D102","D103"]}

[tool.ruff.lint.pydocstyle]
convention = "google"

# Static analysis tools configuration
[tool.pyright]
extraPaths = ["src", "lib"]
pythonVersion = "3.8"
pythonPlatform = "All"

[tool.pytest.ini_options]
minversion = "6.0"
log_cli_level = "INFO"
asyncio_mode = "auto"
addopts = "--tb=native --verbose --capture=no --log-cli-level=INFO"

[tool.codespell]
skip = ".git,.tox,build,venv*"
ignore-words-list = "assertIn"
